Addressing comments for PR
