+++ sigllm/primitives/formatting/value_concatenation.py
@@ -1,6 +1,6 @@
-from .multivariate_formatting import MultivariateFormattingMethod
 import numpy as np
 
+from sigllm.primitives.formatting.multivariate_formatting import MultivariateFormattingMethod
 
 class ValueConcatenation(MultivariateFormattingMethod):
     def __init__(self, verbose: bool = False, **kwargs):
@@ -20,10 +20,3 @@ class ValueConcatenation(MultivariateFormattingMethod):
         ]
         out = np.array(result, dtype=object)
         return out
-
-
-
-if __name__ == "__main__":
-    method = ValueConcatenation()
-    method.test_multivariate_formatting_validity(verbose=False)
-    method.run_pipeline()
